Do not merge individual components, do not import external ontologies.
Do not merge all the individual taxon-specific ontologies into a single
component that is then imported in the `-edit` file. Instead, let's
import all individual components directly in the -edit file. This will
allow users to edit the individual components straight from Protege when
working on the main -edit file (otherwise, any manual change to a
taxon-specific life term from the -edit file would get written to the
`merged-manual.owl` file, which is automatically re-generated from the
indivual files -- in other words, any manual change would get lost).

Also, do not import the externally maintained taxon-specific ontologies
such as FBdv or WBls). Given the new focus of this repository, they are
not needed.

Lastly, we import the life-stages-minimal subset from Uberon, instead of
maintaining a copy of it.