Remove all the `ssso-merged` stuff.
It is no longer in the purview of this repository to provide ssso-merged
and ssso-merged-uberon. The only release products of this project are
now the ODK-generated `life-stages-{full,base}` (containing all the
taxon-specific ontologies maintained as part of this project) and the
individual taxon-specific ontologies themselves, along with the Uberon
mappings.