More cleanup.
All the -bridge-params.tsv files were only needed to generate the
bridges to Uberon, which are no longer the responsibility of this
project.

The -bridge-to-uberon.obo files were the bridges themselves.

And the -uberon.obo files were the merge product of a taxon-specific
ontology and its Uberon bridge.

None of that is needed anymore.