Remove problematic xref to Wikipedia:Gestational_age.
The HsapDv component contains the following typedef:

```
[Typedef]
id: month_of_gestation
...
xref: Wikipedia:Gestational_age
```

The xref was probably intended as a *source* for the definition of
"month of gestation", but the problem is that in the OBO format, a xref
within a Typedef frame is interpreted as the foreign property that the
typedef is an alias for (OBO Flat File Format 1.4, §5.9.3: "If an
Unprefixed-ID is declared in a Typedef frame, and that Typedef frame has
an xref clause, then the value of that xref clause is used in the
translation").

When converting hsapdv.obo to OWL, this leads annotations like this one:

```
property_value: month_of_gestation "3.0" xsd:float
```

to be translated as

```
AnnotationAssertion(Wikipedia:Gestational_age HsapDv:0000197 "3.0"^^xsd:float)
```

which is obviously completely wrong.