Use Uberon's life-stages-core instead of life-stages-minimal.
The life-stages-minimal subset of Uberon is made from a version of the
ontology that contain _only_ is-a and part-of relationships, excluding
any other kind of relationships.

Life stage terms in the original Uberon are connected through a handful
of other relations (e.g. preceded_by), and there should be no reason to
use a version in which those relations have been removed.